advanced/submissions/team-members/prateek-mulye/requirements.txt
langgraph
langchain-openai
langchain-core
langchain-community
langchain-pinecone
langchain-tavily
pinecone-client
gradio
pandas
yfinance
tradingview-ta
python-dotenv
plotly>=5.0.0
